// types.ts
export interface CountryBrief {
  name: string;
  commonName: string;
  flag: string;
  cca2: string;
  mapUrl?: string;
}

export interface CountryDetails {
  name: string;
  continent: string;
  capital: string;
  language: string;
  population: string;
  religion: string;
  area: string;
  currency: string;
  bdtExchangeRate: string;
  isoCode: string;
  dialingCode: string;
  history: string;
  mapUrl?: string;
}